Refactor tenant payment edit modal to use server-side missed payments
